moved all db related code to one file & modified testing ini
