+++ pytest.ini
@@ -1,3 +1,3 @@
 [pytest]
-testpaths=src/tests website/tests
+testpaths=website/tests
 filterwarnings=ignore::DeprecationWarning
